+++ packages/rake-db/src/index.ts
@@ -6,3 +6,4 @@ export * from './migration/migration';
 export { rakeDb } from './rakeDb';
 export * from './ast';
 export type { RakeDbConfig, AppCodeUpdater } from './common';
+export * from './migration/manageMigratedVersions';